Updated to support the x86_64 bit architecture. In each of these files the compiler complained about a loss of precision when ask to cast a void * to an integer. Under the i386 a void * pointer and an integer are both represented by 32 bits, but under the x86_64 architecture, a void * pointer is 64 bits. The solution was to use a cast to long, which works for both architectures.

+++ trunk/src/psos/PsosEventHandler.cpp
@@ -223,7 +223,7 @@ namespace Psos
     MIRO_DBG(PIONEER, LL_PRATTLE,"Psos::EventHandler: handle timer event.");
 
     Miro::Guard guard(synchMutex);
-    switch ((const int)arg) {
+    switch ((const long)arg) {
     case SYNCH_TIMER:
       MIRO_DBG_OSTR(PIONEER, LL_DEBUG, "Synch status: " << synch);
       if (synch < 3) {
