+++ trunk/src/video/VideoDeviceBTTV.cpp
@@ -134,7 +134,7 @@ namespace Video
       if (err == -1)
 	throw Miro::CException(errno, "DeviceBTTV::handleConnect() - VIDIOCGMBUF");
       map_ = (unsigned char*)mmap(0, gb_buffers.size, PROT_READ|PROT_WRITE, MAP_SHARED, ioBuffer_.get_handle(), 0);
-      if ((int)map_ == -1)
+      if ((long)map_ == -1)
 	throw Miro::CException(errno, "DeviceBTTV::handleConnect() - mmap()");
     }
 
@@ -201,7 +201,7 @@ namespace Video
     delete[] channels;
     channels = NULL;
 
-    if ((int)map_ != -1 && map_ != NULL) {
+    if ((long)map_ != -1 && map_ != NULL) {
       munmap(map_, gb_buffers.size);
       map_ = NULL;
     }
